Add errors file
